requirements.txt
python-telegram-bot==20.7
openai-whisper==20231117
ffmpeg-python==0.2.0
